Refactored directories, tweaked members export
This commit refactors the directory structure as explained in README.rst.
It also adds some changes to the script that exports membership info
from the MySQL database. These changes are largely to enable the script
to place files it creates in the right place in the new structure. Calling
of the export.sql file is now automated via a subprocess call.

--HG--
rename : data.py => scripts/data.py
rename : export.sql => scripts/export.sql